backend/requirements.txt

Django>=5.1.4
djangorestframework
django-cors-headers
pillow